Refactor: ProfileUpdate 컴포넌트 분리
